# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Be
  description: |
    The API provides CSR backend API.
basePath: /api/
schemes:
  - https
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /equipment/status_names:
    post:
      summary: Creating a new equipment status.
      security:
        - Bearer: [ ]
      tags:
        - EquipmentStatusName
      operationId: postEquipmentStatusName
      parameters:
        - name: name
          in: body
          description: Status Name for equipment
          required: true
          schema:
            $ref: "#/definitions/NewEquipmentStatusName"
      responses:
        201:
          description: Status name for equipment has been created
          schema:
            $ref: "#/definitions/SuccessEquipmentStatusNameOperationResponse"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/SwaggerError"
    get:
      summary: List all statuses.
      security:
        - Bearer: [ ]
      tags:
        - EquipmentStatusName
      operationId: ListEquipmentStatusNames
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/ListEquipmentStatusNames"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/SwaggerError"
...
definitions:
...
  SwaggerError: # Do not delete it. It is used for generating error responses.
    type: object
    required:
      - code
      - message
    properties:
        code:
            type: integer
            format: int32
        message:
            type: string
        details:
            type: string
...
  EquipmentStatusName:
    type: object
    required:
      - name
    properties:
      id:
        type: integer
      name:
        type: string
  NewEquipmentStatusName:
    type: object
    required:
      - name
    properties:
      name:
        type: string
  ListEquipmentStatusNames:
    type: array
    items:
      $ref: "#/definitions/EquipmentStatusName"
  SuccessEquipmentStatusNameOperationResponse:
    type: object
    required:
      - data
    properties:
      data:
        $ref: "#/definitions/EquipmentStatusName"
...